Reorganized the file structures and modified the run_dev script
