+++ examples/basic/src/hello_world.cc
@@ -1,4 +1,4 @@
-#include "examples/basic/hello_world.h"
+#include "../include/hello_world.h"
 
 void PrintHello() {
     std::cout << "Hello, World!" << std::endl;
