+++ examples/cuda/include/simple_example.h
