+++ examples/cuda/src/simple_example.cu
