+++ examples/test/BUILD.bazel
@@ -1,6 +1,6 @@
 cc_test(
     name = "simple_test",
-    srcs = ["simple_test.cpp"], 
+    srcs = ["src/simple_test.cc"], 
     deps = [
         "@catch2//:catch2",        
         "@catch2//:catch2_main",
